Set the container default before loading source
Problem:
When a user want to send_action to Skygear on bootsrtap, he will try to
instantiate a SkygearContianer instance. However the skygear container
is not yet loaded the default config from environment variable. It make the
developer have to load it from the environment by themselves.

Fix:
Set the container default before attempting to load the plugin source, when
loading the plugin source. Developer got a SkygearContainer with proper
defaults.